refactor(ws-chatbot): WebSocket 경로 및 핸들러 통합 - WebSocket API 라우트 경로를 `/chatbot`으로 통합 - 질문 및 추천 요청 핸들러를 `websocket_endpoint`로 통합 및 리팩토링 - 데이터 구조 및 메시지 타입 체계를 일관되게 수정 (`CREATE_QUESTION`, `RECOMMEND_REQUEST` 등) - 중복된 코드를 제거하고 로깅 메시지 정리 (#61)
